Issues with Level 2
Ran into an issue trying to build after adding background
data for level 2. The specific error I got was this:

ERROR: address overflow (addr 84dd >= 8000)

I believe this has something to do with the ROM being too
large (which is hard to believe considering how small the
game is currently).

I'm going to try transitioning to gbdk-n to hopefully get
around the issues I've been having with the original gbdk.