Move Level 1 to it's own file
Moved level 1 to it's own source file. This is a trend I
will continue with until I find a decent way to store the
level data in a binary/precompiled format.

Also decided to switch to only having one Level object that
is re-used for each level and call the init function when
the level is loaded. It costs a small amount of time for
a huge space savings (hopefully).