Multiple ROM banks and Level Menu
Finally figured out support for multiple ROM banks, which will be
necessary as the game grows since only using two banks would limit
my game to 32k (which is way too small). Currently there are 4 banks
but more will be added as more levels are created.

Also got the backbone of the level menu working. The background and
selector sprite are working, now just need to add the level numbers and
the "select" functionality (press button and level starts).

Other minor changes:
* Changed all background and tile arrays are now const so they are
placed in the ROM instead of RAM.